Regras de negócio para deleção de Autores(não se pode deletar autores que possuem livros)

+++ src/main/java/br/com/caiogit/datajpa/libraryapi/service/AutorService.java
@@ -1,8 +1,11 @@
 package br.com.caiogit.datajpa.libraryapi.service;
 
+import br.com.caiogit.datajpa.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import br.com.caiogit.datajpa.libraryapi.model.Autor;
 import br.com.caiogit.datajpa.libraryapi.repository.AutorRepository;
+import br.com.caiogit.datajpa.libraryapi.repository.LivroRepository;
 import br.com.caiogit.datajpa.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,17 +13,13 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService
 {
 
     private final AutorRepository autorRepository;
+    private final LivroRepository livroRepository;
     private final AutorValidator validator;
-
-    public AutorService(AutorRepository autorRepository, AutorValidator validator)
-    {
-        this.autorRepository = autorRepository;
-        this.validator = validator;
-    }
 
 
     public Autor salvarAutor(Autor autor)
@@ -44,8 +43,17 @@ public class AutorService
         return autorRepository.findById(id);
     }
 
+    private boolean possuiLivro(Autor autor)
+    {
+        return livroRepository.existsByAutor(autor);
+    }
+
     public void deletarAutor(Autor autor)
     {
+        if(possuiLivro(autor))
+        {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um autor que possui livros cadastrados");
+        }
         autorRepository.delete(autor);
     }
